requirements.txt
Django>=2.2.6,<2.3.0
djangorestframework>=3.10.3,<3.11.0
pylint-django>=2.0.12,<2.1.0
mysqlclient>=1.4.6,<1.5.0
django-mysql>=3.3.0,<3.4.0
celery>=4.4.0,<4.5.0
redis>=3.3.11,<3.4.0
elasticsearch-dsl>=6.0.0,<7.0.0
django-elasticsearch-dsl>=6.0.0,<7.0.0
django-elasticsearch-dsl-drf>=0.20.3,<0.21.0
